Rearragnged src folder
